B3Cotacoes/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "b3_cotacoes"
version = "1.0.0"
dependencies = [
    "playwright",
    "pandas",
    "python-dotenv",
    "openpyxl"
]